+++ content-moderation-monitor/scripts/private/CMMClient.psm1
@@ -8,7 +8,7 @@
 
 .NOTES
     Module: CMMClient.psm1
-    Version: 0.1.0
+    Version: 0.2.0
     Author: FSI Agent Governance Team
 #>